# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      senha:
        type: string
    type: object
  auth.TokenResponse:
    properties:
      token:
        type: string
    type: object
  clientes.Cliente:
    properties:
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
      telefone:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: suporte@exemplo.com
    name: Suporte da API
  description: API robusta desenvolvida em Go com Clean Architecture, PostgreSQL e
    JWT.
  termsOfService: http://swagger.io/terms/
  title: API REST - CRUD Go
  version: "1.0"
paths:
  /clientes:
    get:
      description: Retorna uma lista paginada de clientes
      parameters:
      - default: 10
        description: Limite por página
        in: query
        name: limite
        type: integer
      - default: 1
        description: Número da página
        in: query
        name: pagina
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/clientes.Cliente'
            type: array
        "500":
          description: Erro ao buscar clientes
          schema:
            type: string
      summary: Lista todos os clientes
      tags:
      - clientes
    post:
      consumes:
      - application/json
      description: Cria um novo cliente no sistema
      parameters:
      - description: Dados do Cliente
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/clientes.Cliente'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cadastra um novo cliente
      tags:
      - clientes
  /clientes/{id}:
    delete:
      description: Exclui um cliente do sistema com base no ID
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Cliente não encontrado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove um cliente
      tags:
      - clientes
    get:
      description: Retorna os dados de um cliente específico com base no ID
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clientes.Cliente'
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Cliente não encontrado
          schema:
            type: string
      summary: Busca um cliente pelo ID
      tags:
      - clientes
    put:
      consumes:
      - application/json
      description: Altera os dados de um cliente com base no ID
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: integer
      - description: Novos dados do Cliente
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/clientes.Cliente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido ou dados inválidos
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Atualiza um cliente existente
      tags:
      - clientes
  /login:
    post:
      consumes:
      - application/json
      description: Autentica um usuário com e-mail e senha e retorna um token JWT
        de acesso
      parameters:
      - description: Credenciais de Acesso (E-mail e Senha)
        in: body
        name: credenciais
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Dados de login inválidos
          schema:
            type: string
        "401":
          description: E-mail ou senha incorretos
          schema:
            type: string
      summary: Realiza a autenticação do usuário
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"